techDebt(registrationRefactor) Moved anonymous .post function into registerAccount function and into a new file
